finish restructuring
