+++ frontend/src/app/global/styles/globals.css
