+++ frontend/src/app/layout.tsx
@@ -1,4 +1,4 @@
-import "./styles/globals.css";
+import "@global/styles/globals.css";
 import { AuthProvider } from "./global/AuthContext";
 
 export default function RootLayout({ children }: { children: React.ReactNode }) {
