+++ frontend/src/app/protected/layout.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import Header from "@/lib/components/Header";
+import Header from "@lib/components/ui/Header";
 
 export default function ProtectedLayout({ children }: { children: React.ReactNode }) {
   return (
