+++ frontend/src/lib/components/Character/character.ts
