+++ frontend/src/lib/components/Character/sampleCharacter.json
