+++ frontend/src/lib/components/Character/sampleCharacter.ts
