+++ frontend/src/lib/components/Character/sampleCharacter_dynamodb.json
