+++ frontend/src/lib/components/Footer.tsx
