+++ frontend/src/lib/components/ui/Footer.tsx
@@ -1,4 +1,4 @@
-import "../styles/globals.css";
+import "@global/styles/globals.css";
 
 export default function Footer() {
   return (
