+++ frontend/src/lib/components/ui/Header.tsx
@@ -1,9 +1,8 @@
 import Link from "next/link";
 // import { useRouter } from 'next/navigation'
 
-import { Button } from "@/lib/components/ui/button";
-
-import "../../src/app/styles/globals.css";
+import { Button } from "@lib/components/ui/button";
+import "@global/styles/globals.css";
 
 export default function Header() {
   //const { logout } = useAuth()
